requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.3
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
captum==0.5.0
    # via -r requirements.in
certifi==2022.9.24
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via wandb
contourpy==1.0.6
    # via matplotlib
cycler==0.11.0
    # via matplotlib
docker-pycreds==0.4.0
    # via wandb
fire==0.4.0
    # via lightning-utilities
fonttools==4.38.0
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.11.0
    # via pytorch-lightning
gitdb==4.0.10
    # via gitpython
gitpython==3.1.29
    # via wandb
idna==3.4
    # via
    #   requests
    #   yarl
joblib==1.2.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
lightning-utilities==0.3.0
    # via pytorch-lightning
matplotlib==3.6.2
    # via captum
monai==1.0.1
    # via -r requirements.in
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.23.5
    # via
    #   -r requirements.in
    #   captum
    #   contourpy
    #   matplotlib
    #   monai
    #   pandas
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   tensorboardx
    #   torchmetrics
    #   torchvision
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
packaging==21.3
    # via
    #   matplotlib
    #   pytorch-lightning
    #   torchmetrics
pandas==1.5.2
    # via -r requirements.in
pathtools==0.1.2
    # via wandb
pillow==9.3.0
    # via
    #   matplotlib
    #   torchvision
promise==2.3
    # via wandb
protobuf==3.20.1
    # via
    #   tensorboardx
    #   wandb
psutil==5.9.4
    # via wandb
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==1.8.3.post1
    # via -r requirements.in
pytz==2022.6
    # via pandas
pyyaml==6.0
    # via
    #   pytorch-lightning
    #   wandb
requests==2.28.1
    # via
    #   fsspec
    #   torchvision
    #   wandb
scikit-learn==1.2.0
    # via -r requirements.in
scipy==1.9.3
    # via scikit-learn
sentry-sdk==1.12.0
    # via wandb
setproctitle==1.3.2
    # via wandb
shortuuid==1.0.11
    # via wandb
six==1.16.0
    # via
    #   docker-pycreds
    #   fire
    #   promise
    #   python-dateutil
    #   wandb
smmap==5.0.0
    # via gitdb
tensorboardx==2.5.1
    # via pytorch-lightning
termcolor==2.1.1
    # via fire
threadpoolctl==3.1.0
    # via scikit-learn
torch==1.13.0
    # via
    #   -r requirements.in
    #   captum
    #   monai
    #   pytorch-lightning
    #   torchaudio
    #   torchmetrics
    #   torchvision
torchaudio==0.13.0
    # via -r requirements.in
torchmetrics==0.11.0
    # via
    #   -r requirements.in
    #   pytorch-lightning
torchvision==0.14.0
    # via -r requirements.in
tqdm==4.64.1
    # via pytorch-lightning
typing-extensions==4.4.0
    # via
    #   pytorch-lightning
    #   torch
    #   torchvision
urllib3==1.26.13
    # via
    #   requests
    #   sentry-sdk
wandb==0.13.5
    # via -r requirements.in
wheel==0.38.4
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-runtime-cu11
yarl==1.8.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
